Convert drawing code to double from int, and remove some unneeded 'floor' commands. Results in a curve drawing speedup of ~40% on intel core.
Will also allow antialiased lines.  We'll test that next.



git-svn-id: svn://anonsvn.kde.org/home/kde/branches/work/kst/portto4/kst@1073877 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ src/libkstmath/curvepointsymbol.h
@@ -27,7 +27,7 @@ namespace Kst {
 
 namespace  CurvePointSymbol {
     /** draw the point on a painter the scale of the point is based on size */
-    KST_EXPORT void draw(int type, QPainter *p, int x, int y, int lineSize = 0, int size = -1);
+    KST_EXPORT void draw(int type, QPainter *p, double x, double y, int lineSize = 0, int size = -1);
     /** Get the dimension of a point */
     KST_EXPORT int dim(const QRect &window);
 }
